backend/appbase/requirements.txt
django
django-rest-framework
djangorestframework-simplejwt
django-cors-headers
